+++ src/lint/rules/tests/common.rs
@@ -24,17 +24,3 @@ pub fn set_up() -> CurrentRules {
     let cfg = LintCfg::default();
     instantiate_rules(&cfg)
 }
-
-pub fn assert_indentation(
-    code: &str, expected_errors: usize, rules: CurrentRules)
-{
-    let lint_errors = run_linter(code, &rules);
-    let Ok(ref lint_errors) = lint_errors else {
-        panic!();
-    };
-    let mut indent_errors: Vec<&LocalDMLError> = vec!();
-    for error in lint_errors {
-        indent_errors.push(error);
-    }
-    assert_eq!(indent_errors.len(), expected_errors, "{:#?}", lint_errors);
-}
